Created handlers folder and repository folder - Modified client.js /register endpoint to only call methods from the handler and the repository
